// frontend/app/src/app/core/api/model/analyticsSummaryResponse.ts
/**
 * FS Bus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AnalyticsVehicleScoreResponse } from './analyticsVehicleScoreResponse';
import { AnalyticsTopKpiResponse } from './analyticsTopKpiResponse';
import { AnalyticsGaugeScoreResponse } from './analyticsGaugeScoreResponse';
import { AnalyticsLastEventResponse } from './analyticsLastEventResponse';
import { AnalyticsVehiclePerformanceResponse } from './analyticsVehiclePerformanceResponse';
import { AnalyticsMetricTileResponse } from './analyticsMetricTileResponse';


export interface AnalyticsSummaryResponse { 
    /**
     * UTC timestamp for this summary.
     */
    generated_at: string;
    source_date_from?: string | null;
    source_date_to?: string | null;
    top_kpis: Array<AnalyticsTopKpiResponse>;
    gauge_scores: Array<AnalyticsGaugeScoreResponse>;
    metric_tiles: Array<AnalyticsMetricTileResponse>;
    last_events: Array<AnalyticsLastEventResponse>;
    vehicle_performance: Array<AnalyticsVehiclePerformanceResponse>;
    vehicle_scores: Array<AnalyticsVehicleScoreResponse>;
}