// frontend/app/src/app/core/api/model/analyticsVehicleScoreResponse.ts
/**
 * FS Bus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AnalyticsVehicleScoreResponse { 
    fleet_no: string;
    registration: string;
    operator: string;
    score: number;
}